+++ lib/screen/constants.dart
@@ -3,4 +3,4 @@ import 'package:flutter/material.dart';
 const kTextColor = Colors.white;
 const kTextLightColor = Colors.black;
 
-const kDefaultPaddin = 10.0;
+const kDefaultPaddin = 8.0;
